refactor: stock-detail 컴포넌트 경로 이동 및 전체 css grow -> flex-1로 변경
