+++ src/components/orders/stock-detail/StockDetailTabs.tsx
